game-03: Copy moddable-platformer to game-03
Copy moddable-platformer to game-03, but detach GDscripts basically.
However, to make player move and collect coins, add some Block Coding
blocks to implement the behaviors. Besides, the player uses A & D keys
to walk left & right and W key for jumping.

This is not a beginning step as the baseline. More features will come
later.

https://phabricator.endlessm.com/T35632