game-03: Only raise_flag & show_win when the player reaches the flag & collects >= 19 coins
Add a coins variable in the flag's BlockCode node and add the flag node
into HUD group. Define collect_a_coin method which increases the coins
variable, just like the simple_score's BlockCode node. It will be
invoked when the player enters a coin, too.

Then, only raise_flag and show_win when the player reaches the flag and
collects more, or equal 19 coins. Here are 19 coins in the game.

https://phabricator.endlessm.com/T35632