Stop music when game ends
The BackgroundMusic node should have its Autoplay property set to start the
music when the game starts. I haven't done this because it's so annoying in
development.

We have no blocks for AudioStreamPlayer nodes so we can't call stop on the
track, but we can just remove the player from the scene.