Port game-01 to current plugin snapshot
To get this working I rebuilt the 3 block code programs with exactly the same
blocks as before.

I then added a CollisionPolygon2D to the SimpleCharacter. This is because in
the old plugin version a 100×100 square CollisionShape2D was unconditionally
added; whereas in the new version this is only added if the SimpleCharacter's
texture property is set. But the texture property is not set.

We could instead set the texture and drop the Sprite2D2 child. However it's not
as trivial as setting the texture property because the sprite is the wrong size
and is scaled down to 0.321 its full size by the Sprite2D2.

So, add a CollisionPolygon2D, and adjust the position of the dead zone
accordingly.