Add web export preset
This preset is just the default settings for the Web target (as of
v4.3.stable.flathub [77dcf97d8], at least).

The important line is:

variant/thread_support=false

This selects a build of the Godot web runtime that does not use threads; and
hence does not use SharedArrayBuffer; and hence can be used on GitHub Pages
where the necessary headers to allow SharedArrayBuffer to be used cannot be
set.

https://phabricator.endlessm.com/T35699